move new base conversion to new file
